+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## {{releaseVersion}} - {{releaseDate}}
 
+### Fixed
+
+- Swift-testing test runs are marked as 'started' in the UI immediately, not after compilation finishes ([#2079](https://github.com/swiftlang/vscode-swift/pull/2079))
+
 ## 2.16.1 - 2026-02-02
 
 ### Fixed
 
 - Fixed extension activation issues involving swiftly ([#2066](https://github.com/swiftlang/vscode-swift/pull/2066))
+- Fixed not all settings notifying the user when malformed ([#2086](https://github.com/swiftlang/vscode-swift/pull/2086))
 
 ## 2.16.0 - 2026-01-28